+++ src/conformance/mod.rs
@@ -1,3 +1,5 @@
+//! The conformance feature is not yet up to spec. Usage is not yet recommended.
+
 mod auth;
 mod operation;
 mod param;
@@ -6,10 +8,4 @@ mod response;
 mod runner;
 mod test;
 
-pub use auth::*;
-pub use operation::*;
-pub use param::*;
-pub use request::*;
-pub use response::*;
-pub use runner::*;
-pub use test::*;
+pub use self::{auth::*, operation::*, param::*, request::*, response::*, runner::*, test::*};