+++ CHANGELOG.md
@@ -2,7 +2,11 @@
 
 ## Unreleased
 
-- The `spec::Parameter::schema` field is now of type `ObjectOrReference<Schema>`.
+- Rename `spec::{Schema => ObjectSchema}` struct.
+- Add `spec::BooleanSchema` struct.
+- Add `spec::Schema` enum.
+- The `spec::ObjectSchema::addition_properties` field is now of type `Option<Schema>`.
+- The `spec::Parameter::schema` field is now of type `ObjectOrReference<ObjectSchema>`.
 - Add `Operation::extensions` field.
 - Minimum supported Rust version (MSRV) is now 1.70.